// src/modules/patients/interfaces/record.interface.ts
import { Patient } from "src/modules/admin/interface/patient.interface";

export interface Record {
    id?: string;
    startDate?: Date;
    user?: Patient;
    age?: number;
    startWeigth?: number;
    gender?: string;
    bdate?: Date;
    civilStatus?: string;
    religion?: string;
    scolarship?: string;
    occupation?: string;
    address?: string;
    phone?: string;
    diarrhea?: boolean;
    constipation?: boolean;
    gastritis?: boolean;
    ulcers?: boolean;
    nausea?: boolean;
    pyrosis?: boolean;
    vomiting?: boolean;
    colitis?: boolean;
    stainedTeeth?: boolean;
    headache?: boolean;
    kneePain?: boolean;
    anemia?: boolean;
    diabetes?: boolean;
    has?: boolean;
    overweightObesity?: boolean;
    surgeries?: boolean;
    otherDiseases?: string;
    observations?: string;
    diagnosedDisease?: string;
    medications?: boolean;
    whichMedications?: string;
    supplements?: string;
    familyObesity?: boolean;
    familyDiabetes?: boolean;
    familyHta?: boolean;
    familyCancer?: boolean;
    familyOther?: string;
    physicalActivity?: string;
    exercise?: boolean;
    exerciseType?: string;
    exerciseFrequency?: string;
    exerciseDuration?: string;
    exerciseSince?: string;
    alcohol?: string;
    tobacco?: string;
    coffee?: string;
    otherSubstances?: string;
    generalSigns?: string;
    mealsPerDay?: number | null;
    breakfastTime?: string;
    lunchTime?: string;
    dinnerTime?: string;
    eatsBetweenMeals?: boolean;
    whatEatsBetweenMeals?: string;
    whoPreparesFood?: string;
    appetite?: string;
    hungryTime?: string;
    foodDiscomfort?: string;
    foodAllergy?: boolean;
    foodAllergySpecific?: string;
    addsSalt?: boolean;
    addsSaltReason?: string;
    cookingFat?: string;
    nutritionalGuidance?: boolean;
    nutritionalGuidanceReason?: string;
    weightLossMedication?: boolean;
    weightLossMedicationDetails?: string;
    waterIntake?: number | null;
    isEdit?: boolean;
}